requirements.txt
aiofiles
aiohttp
asyncio
pyrogram==1.4.16
git+https://github.com/MrAdityaXD/yt-dl@master
python-arq
python-dotenv
py-tgcalls
requests
TgCrypto
wget
youtube-search-python==1.4.6
